SpellcheckerService2/requirements.txt
pyspellchecker
pydantic
uvicorn
fastapi
requests